lekinetworks.bot/requirements.txt
httpx
aiogram
python-decouple
python-dotenv